Lots of bug fixes - Doubles removed for longs to store money - Serialization functions fixed where needed - UI bugs removed - Main.cpp removed

+++ Sockets/Serialization/Serialize.h
@@ -7,13 +7,13 @@
 #define BANKINGAPPSERVER_SERIALIZE_H
 
 std::string serializecredentialcreate(ui::Credential* cred);
-std::string serializewithdrawl(ui::Credential* cred,double amount);
-std::string serializedeposit(ui::Credential* cred,double amount);
+std::string serializewithdrawl(ui::Credential* cred,long amount);
+std::string serializedeposit(ui::Credential* cred,long amount);
 int determineprocess(char* input);
 std::string serializecredentiallogin(ui::Credential* cred);
 bool determinesuccess(char* input);
-double returnbalance(char* input);
-double serverreturnamount(char* input);
+long returnbalance(char* input);
+long serverreturnamount(char* input);
 std::string serverreturnpassword(char* input);
 std::string serverreturnusername(char* input);
 #endif //BANKINGAPPSERVER_SERIALIZE_H
